+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](http://semver.org).
 - Cosmetic - Fixes and adjustments to README and other docs
 - Added - cli.beforeExec() and cli.afterExec() for code common to all commands
 - Added - cli.parseAborted() to test whether cli.parseExit() was called
+- Added - cli.print*() overloads that append to a string
 
 ## dimcli 7.3.0 (2024-06-19)
 - Fixed - Assert calling cli.title() immediately after creating new option